+++ Changelog.md
@@ -5,8 +5,9 @@ __New Features__
   - Replaces `PortableHeater` and `FixedHeater` with `SpaceHeater`.
 - Adds manufactured home belly as a foundation type and allows modeling ducts in a manufactured home belly.
 - Output updates:
-  - ReportUtilityBills measure: Allows reporting monthly utility bills in addition to (or instead of) annual bills.
+  - Adds "Peak Electricity: Annual Total (W)" output.
   - Adds battery resilience hours output; allows requesting timeseries output.
+  - ReportUtilityBills measure: Allows reporting monthly utility bills in addition to (or instead of) annual bills.
 
 __Bugfixes__
 - Fixes lighting multipliers not being applied when kWh/yr inputs are used.